+++ lib/rpmsg_lite/porting/environment/rpmsg_env_freertos.c
@@ -549,7 +549,7 @@ void env_release_sync_lock(void *lock)
  *
  * Suspends the calling thread for given time , in msecs.
  */
-void env_sleep_msec(uint32_t num_msec)
+void env_sleep_msec(uintptr_t num_msec)
 {
     vTaskDelay(num_msec / portTICK_PERIOD_MS);
 }
