+++ lib/rpmsg_lite/porting/environment/rpmsg_env_threadx.c
@@ -481,7 +481,7 @@ void env_release_sync_lock(void *lock)
  *
  * Suspends the calling thread for given time , in msecs.
  */
-void env_sleep_msec(uint32_t num_msec)
+void env_sleep_msec(uintptr_t num_msec)
 {
     tx_thread_sleep(num_msec);
 }
